AI-NLP/keyframe_extractor_classifier/requirements.txt
appdirs==1.4.4
APScheduler==3.9.1.post1
asgiref==3.4.1
astroid==2.6.6
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
blis==0.7.4
boto3==1.24.68
botocore==1.27.68
bson==0.5.10
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
cymem==2.0.5
distlib==0.3.3
Django==3.2.5
django-apscheduler==0.6.0
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
filelock==3.1.0
glob2==0.7
greenlet==2.0.1
gunicorn==20.1.0
huggingface-hub==0.0.19
identify==2.2.15
idna==3.2
ImageHash==4.2.1
imageio==2.9.0
importlib-metadata==1.3.0
iniconfig==1.1.1
jmespath==1.0.1
joblib==1.1.0
jwt==1.3.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
murmurhash==1.0.5
mypy-extensions==0.4.3
mysqlclient==2.0.3
networkx==2.6.3
nodeenv==1.6.0
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.3
passlib==1.7.4
pathspec==0.9.0
pathy==0.6.0
Pillow==8.3.2
plac==1.1.3
platformdirs==2.4.0
pluggy==0.13.1
pre-commit==2.15.0
preshed==3.0.5
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
s3transfer==0.6.0
sacremoses==0.0.46
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
spacy==2.3.5
spacy-legacy==3.0.8
SQLAlchemy==1.4.44
sqlparse==0.4.2
srsly==1.0.5
thinc==7.4.5
tifffile==2021.8.30
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.1
torch==1.9.1
tqdm==4.62.2
transformers==4.11.3
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.10.0.2
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.6
virtualenv==20.8.1
wasabi==0.8.2
wrapt==1.12.1
zipp==3.10.0
